started adding calculations for prizes
